[gldemo] factor out glkit module
